Implement modules exports
